Updating shading test
Added some extra denting and an optional darkening along with it to the metal shader.
The darkening and denting also has some effect on the paint layers.
Added some controls to the dirt shader as well. It's alpha is controlling, how much of it is visible, so it can be colored independently.
Sample maps are packed into the blend file.

Needs a lot of tidying up, and support for the decals/transparency as well.