Create pioneer-preview.blend
Adding a simple preview material that mimics how the model would look in-game with the current rendering. Accepts the textures the game accepts: Diff, Spec, Glow, Norm, Pattern but not AO.
Works with both Eevee and Cycles.
Pattern colors are adjustable.
Has a transparency slider as well (eg: for cockpits). Note: transparency needs to be enabled per material for Eevee.
Should work for the manufacturer decals if their alpha is connected to the transparency input.