Create sinonatrix_redesign.blend
The model file for the redesigned Sinonatrix, based on the model by Torham2234, who modeled it based on my simple sketch.
It is used as a testbed for some design language changes, and for the PBR&decals workflow. Some decals on it are from the decalMachine addon for now. It might not find those, but should find the RCS one, if the decals/OPLI directory is added to the addon as a library. It should also be able to find it, if you pull this repository.